AG-13291 Fix tabindex>0 tab-guards
Also, disable tab guards tabindex=0. To properly support this like in AG Grid
we'll need to set the tabindices to -1 on all the focusable elements in the
chart and manage Tab key inputs ourselves. This isn't going to happen any time
soon, therefore just disable tab guards to avoid breaking the default behaviour.